Further UI refacotring - stage 1
