+++ src/structs.rs
@@ -48,8 +48,9 @@ pub struct GameWindow {
 impl GameWindow {
 
     pub fn new(window_width: f64, window_height: f64, window: PistonWindow) -> Self {
-        GameWindow { width: window_width, height: window_height,
-            window: Rc::new(RefCell::new(window)) }
+        GameWindow { width: window_width,
+                    height: window_height,
+                    window: Rc::new(RefCell::new(window)) }
     }
 
     #[inline]
