+++ src/ui/window/confirm/mod.rs
@@ -2,13 +2,17 @@
 
 use super::{ActiveWindow, WindowBase};
 
+use piston_window::Context;
+use opengl_graphics::GlGraphics;
+
+
 pub struct ConfirmationWindow {
     state: isize,
 }
 
 impl WindowBase for ConfirmationWindow {
 
-    fn paint(&self) {
+    fn paint(&mut self, c: Context, g: &mut GlGraphics) {
 
     }
